requirements.txt
jupyterlab
matplotlib
numpy
pandas
seaborn
scikit-learn
scipy
statsmodels
pyaudio
librosa
torch
torchaudio
speechbrain
pyannote.core
pyannote.metrics
pyannote.audio
diart
PyQt5